+++ tests/FunctionalTest/ContainerAwareTest.php
@@ -26,6 +26,7 @@ declare(strict_types=1);
 
 namespace Teknoo\Tests\DI\SymfonyBridge\FunctionalTest;
 
+use DI\Definition\Source\SourceCache;
 use Teknoo\Tests\DI\SymfonyBridge\FunctionalTest\Fixtures\ContainerAwareController;
 
 /**
@@ -36,9 +37,41 @@ class ContainerAwareTest extends AbstractFunctionalTest
     /**
      * @link https://github.com/PHP-DI/Symfony-Bridge/issues/2
      */
-    public function testContainerAware()
+    public function testContainerAwareWithoutCacheAndWithoutCompilation()
     {
-        $kernel = $this->createKernel();
+        $kernel = $this->createKernel('empty.yml');
+        $container = $kernel->getContainer();
+
+        /** @var ContainerAwareController $class */
+        $class = $container->get(ContainerAwareController::class);
+
+        self::assertSame($container, $class->container);
+    }
+
+    /**
+     * @link https://github.com/PHP-DI/Symfony-Bridge/issues/2
+     */
+    public function testContainerAwareWithCacheAndWithoutCompilation()
+    {
+        if (!SourceCache::isSupported()) {
+            self::markTestSkipped('APCu is not enabled');
+        }
+
+        $kernel = $this->createKernel('empty_with_cache.yml');
+        $container = $kernel->getContainer();
+
+        /** @var ContainerAwareController $class */
+        $class = $container->get(ContainerAwareController::class);
+
+        self::assertSame($container, $class->container);
+    }
+
+    /**
+     * @link https://github.com/PHP-DI/Symfony-Bridge/issues/2
+     */
+    public function testContainerAwareWithoutCacheAndWithCompilation()
+    {
+        $kernel = $this->createKernel('empty_with_compilation.yml');
         $container = $kernel->getContainer();
 
         /** @var ContainerAwareController $class */